Add reset support in bananapif3.
This commit introduces support for reset functionality on the Banana Pi F3 board.
The necessary code changes are made to initialize and manage reset signals, 
ensuring the correct handling of the board's reset behavior.

[internal message]:
currently i'm stuggling while testing on real hardware,
the build log is here:
https://privatebin.net/?b5c89ed0df1586d1#Qy7gHVYW2uZ1vhVkFkyHsXXSfJ7WpmAY62N7wbpRBPj
and can see the spacemit reset driver is successfully compiled,
but when type `dm tree` the reset driver is not probed by the system:
https://privatebin.net/?0f40b3bdfed9a704#B12WQN45e9Hyma3faZSppFADHJ4amFNw6hfzRYg6MbDV
looking forward some debug tricks or ideas.

Signed-off-by: Huan Zhou <pericycle.cc@gmail.com>

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 1,
    "change-id": "20241202-k1-reset-efa22f5233da",
    "prefixes": []
  }
}